models.py contents moved to models folder
